momentum-add/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.0
Flask==3.0.0
frozenlist==1.4.0
functions-framework==3.5.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud==0.34.0
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-firestore==2.13.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pandas==2.1.4
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
yarl==1.9.4
